Refactor the revoke/delete user api names.
Previously we had the two methods `denyUser` and `removeUser`. `denyUser`
removed the user from the system and `removeUser` revoked access privileges
(while still keeping the user in the db (the idea was probably remove access
privileges)).

Now these functions are called `deleteUser` for complete removal from the
system and `revokeUser` for revoking privileges.